Rename classes and files for file sources
